+++ Libft/ft_memcpy.c
@@ -6,13 +6,13 @@
 /*   By: piotroff <piotroff@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/03 16:45:23 by apiotrau          #+#    #+#             */
-/*   Updated: 2023/05/11 19:14:36 by piotroff         ###   ########.fr       */
+/*   Updated: 2023/05/15 18:30:00 by piotroff         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	*ft_memcpy(void	*dest, const void	*src, unsigned int	n)
+void	*ft_memcpy(void	*dest, const void	*src, size_t	n)
 {
 	char	*d; 
 	char	*s;
