+++ Libft/ft_strlen.c
@@ -6,13 +6,13 @@
 /*   By: piotroff <piotroff@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/02 12:18:20 by apiotrau          #+#    #+#             */
-/*   Updated: 2023/05/11 19:31:58 by piotroff         ###   ########.fr       */
+/*   Updated: 2023/05/15 18:28:35 by piotroff         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int ft_strlen(const char *c)
+size_t ft_strlen(const char *c)
 {
     int i;
