+++ Libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: piotroff <piotroff@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/02 11:14:31 by apiotrau          #+#    #+#             */
-/*   Updated: 2023/05/15 17:49:00 by piotroff         ###   ########.fr       */
+/*   Updated: 2023/05/15 18:30:49 by piotroff         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,25 +26,25 @@ typedef struct s_list
 }	t_list;
 
 int				ft_isalpha(int c);
-int				ft_strlen(const char *c);
+size_t ft_strlen(const char *c);
 int				ft_isprint(int c);
 int				ft_isdigit(int c);
 int				ft_isascii(int c);
 int				ft_isalnum(int c);
 int				ft_toupper(int c);
 int				ft_tolower(int	c);
-int				ft_strncmp(const char *s1, const char *s2, unsigned int n);
-int				ft_memcmp(const void *s1, const void *s2, unsigned int n);
+int	ft_strncmp(const char *s1, const char *s2, size_t n);
+int	ft_memcmp(const void *s1, const void *s2, size_t n);
 int				ft_atoi(const char *nptr);
-unsigned int	ft_strlcpy(char *dest, const char *src, unsigned int size);
-unsigned int	ft_strlcat(char *dest, const char *src, unsigned int size);
-void	*ft_memmove(void *dest, const void *src, unsigned int n);
-void 	*ft_memchr(const void *s, int c, unsigned int n);
-void    *ft_memset(void *s, int c, unsigned int n);
-void    ft_bzero(void *s, unsigned int n);
-void	*ft_memcpy(void	*dest, const void	*src, unsigned int	n);
-void	*ft_calloc(unsigned int nmemb, unsigned int size);
-void	*ft_memccpy(void *dest, const void *src, int c, unsigned int n);
+size_t	ft_strlcpy(char *dest, const char *src, size_t size);
+size_t	ft_strlcat(char *dest, const char *src, size_t size);
+void	*ft_memmove(void *dest, const void *src, size_t n);
+void *ft_memchr(const void *s, int c, size_t n);
+void    *ft_memset(void *s, int c, size_t n);
+void    ft_bzero(void *s, size_t n);
+void	*ft_memcpy(void	*dest, const void	*src, size_t	n);
+void *ft_calloc(size_t nmemb, size_t size);
+void *ft_memccpy(void *dest, const void *src, int c, size_t n);
 void	ft_striteri(char *s, void (*f)(unsigned int, char *));
 void	ft_putchar_fd(char c, int fd);
 void	ft_putstr_fd(char *s, int fd);
@@ -52,12 +52,12 @@ void	ft_putendl_fd(char *s, int fd);
 void	ft_putnbr_fd(int n , int fd);
 char		*ft_strchr(const char *s, int c);
 char		*ft_strrchr(const char *s, int c);
-char		*ft_strnstr(const char *big, const char *little, unsigned int len);
+char *ft_strnstr(const char *big, const char *little, size_t len);
 char		*ft_strdup(const char *src);
 char		*ft_itoa(int n);
 char		*ft_strjoin(char const *s1, char const *s2);
 char		**ft_split(char const *s, char c);
-char		*ft_substr(char const *s, unsigned int start, unsigned int len);
+char *ft_substr(char const *s, unsigned int start, size_t len);
 char		*ft_strtrim(char const *s1, char const *set);
 char		*ft_strmapi(char const *s, char (*f)(unsigned int, char));
 t_list	*ft_lstnew(void *content);
